\name{inspect}
\alias{inspect}
\title{A function to inspect the variables and data you are about to analyse.}
\usage{
inspect()
}
\description{
Use this function to inspect the dataframe and variable you are about to use in your analysis.
Format: inspect(MYDATAFRAME, MYDATAFRAME$MYVARIABLE)}
\examples{
inspect(mpg,mpg$cyl)
}
